add productDetail component
